complete homework 3

+++ src/main/java/co/istad/longfou/restapi/domain/Coffee.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+
+import java.math.BigDecimal;
 //import lombok.Data;
 
 @Getter
@@ -15,7 +17,8 @@ public class Coffee {
     private Integer id;
     private String name;
     private String description;
-    //add price atritute
-    private Double price;
+    //add price attribute
+    //private Double price;
+    private BigDecimal price;
 
 }
